chore: Refactor command loading in handler
The code changes refactor the way commands are loaded in the handler module. Instead of manually importing each command file, the code now dynamically loads all command files in the specified directory. This improves maintainability and allows for easier addition of new commands in the future.

Note: The recent user commits and repository commits do not provide any additional context for the commit message.
